backend/pyproject.toml
[project]
name = "adk-agui-backend"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "sse-starlette>=2.1.3",
    "google-adk>=1.16.0",
    "litellm>=1.43.14",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.1"
]

[tool.uvicorn]
factory = false
reload = true
port = 8080
host = "0.0.0.0"
